deploy: move config change controller to new location
