+++ pkg/deploy/controller/generictrigger/controller_test.go
@@ -1,4 +1,4 @@
-package configchange
+package generictrigger
 
 import (
 	"testing"
